Refactor `pyklatchat_utils` to `klatchat_utils` for consistent naming with observer module
