+++ .github/workflows/unit_tests.yml
@@ -12,6 +12,7 @@ jobs:
     strategy:
       matrix:
         python-version: [ 3.9, '3.10', '3.11', '3.12' ]
+    timeout-minutes: 30
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
