+++ klatchat_utils/common.py
