+++ klatchat_utils/database_utils/mongo_utils/__init__.py
@@ -26,4 +26,10 @@
 # NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 # SOFTWARE,  EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from pyklatchat_utils.database_utils.db_controller import DatabaseController
+from klatchat_utils.database_utils.mongo_utils.structures import (
+    MongoFilter,
+    MongoCommands,
+    MongoQuery,
+    MongoDocuments,
+    MongoLogicalOperators,
+)
