+++ klatchat_utils/database_utils/mongo_utils/queries/dao/personas.py
@@ -25,11 +25,13 @@
 # LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
 # NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 # SOFTWARE,  EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-from pyklatchat_utils.database_utils.mongo_utils.structures import (
-    MongoFilter,
-    MongoCommands,
-    MongoQuery,
+from klatchat_utils.database_utils.mongo_utils import (
     MongoDocuments,
-    MongoLogicalOperators,
 )
+from klatchat_utils.database_utils.mongo_utils.queries.dao.abc import MongoDocumentDAO
+
+
+class PersonasDAO(MongoDocumentDAO):
+    @property
+    def document(self):
+        return MongoDocuments.PERSONAS
