+++ klatchat_utils/database_utils/mongo_utils/queries/dao/prompts.py
@@ -30,13 +30,13 @@ from typing import List
 
 import pymongo
 
-from pyklatchat_utils.database_utils.mongo_utils import (
+from klatchat_utils.database_utils.mongo_utils import (
     MongoDocuments,
     MongoCommands,
     MongoFilter,
     MongoLogicalOperators,
 )
-from pyklatchat_utils.database_utils.mongo_utils.queries.dao.abc import MongoDocumentDAO
+from klatchat_utils.database_utils.mongo_utils.queries.dao.abc import MongoDocumentDAO
 from neon_utils.logger import LOG
