+++ klatchat_utils/database_utils/mongo_utils/structures.py
@@ -29,7 +29,7 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import Any, List, Union
 
-from pyklatchat_utils.common import deep_merge
+from klatchat_utils.common import deep_merge
 
 
 class MongoCommands(Enum):
