+++ klatchat_utils/templates/components/conversation_skins/prompts.html
