+++ klatchat_utils/templates/components/modals/license.html
