+++ klatchat_utils/templates/components/prompt_participant.html
