+++ klatchat_utils/templates/components/prompt_table.html
