+++ klatchat_utils/templates/components/suggestion.html
