+++ klatchat_utils/templates/components/user_message.html
