+++ klatchat_utils/templates/components/user_message_minified.html
