+++ tests/test_connection_utils.py
@@ -31,5 +31,5 @@ from unittest import TestCase
 
 class TestConnectionUtils(TestCase):
     def test_create_ssh_tunnel(self):
-        from pyklatchat_utils.connection_utils import create_ssh_tunnel
+        from klatchat_utils.connection_utils import create_ssh_tunnel
         # TODO
