+++ tests/test_exceptions.py
@@ -31,7 +31,7 @@ from unittest import TestCase
 
 class TestExceptions(TestCase):
     def test_exceptions(self):
-        from pyklatchat_utils.exceptions import KlatchatException, KlatAPIAuthorizationError, MalformedConfigurationException
+        from klatchat_utils.exceptions import KlatchatException, KlatAPIAuthorizationError, MalformedConfigurationException
         self.assertIsInstance(KlatchatException(), KlatchatException)
         self.assertIsInstance(KlatAPIAuthorizationError(), KlatchatException)
         self.assertIsInstance(MalformedConfigurationException(), KlatchatException)
